feat: 후원사 컴포넌트 공통화 및 main page 적용
