+++ src/components/Supports/Supports.tsx
@@ -1,22 +1,24 @@
 import Image from 'next/image';
-import { css, Theme } from '@emotion/react';
+import { css } from '@emotion/react';
 
-import { SupportThumbnail } from '~/components/Supports/SupportThumbnail';
 import { SUPPORTS } from '~/constant/supports';
+import { useCheckWindowSize } from '~/hooks/useCheckWindowSize';
 import { mediaQuery } from '~/styles/media';
 import { theme } from '~/styles/theme';
 
-import { SectionTitleV2 } from '../SectionTitleV2';
+import { SupportThumbnail } from './SupportThumbnail';
 
-export function Supports() {
+export const Supports = () => {
+  const { isTargetSize: isMobileSize } = useCheckWindowSize('mobile');
+
   return (
     <div css={[layoutCss]}>
-      <div css={titleContainerCss}>
-        <SectionTitleV2 css={titleContainerCss}>
-          <span css={titleCss}>후원사</span>
-        </SectionTitleV2>
-        <p css={subTitleCss}>디프만 후원사는 지속해서 추가 예정입니다.</p>
-      </div>
+      <h1 css={introCss.headline}>후원사</h1>
+      <p css={introCss.description}>
+        디프만은 IT비영리단체로 후원을 통해
+        {isMobileSize && <br />}더 많은 교육 기회에 도움을 받고 있습니다.
+      </p>
+
       <ul css={supportContainerCss}>
         {SUPPORTS.map(support => (
           <SupportThumbnail key={support.title} {...support} />
@@ -27,23 +29,42 @@ export function Supports() {
       </ul>
     </div>
   );
-}
+};
 
 const layoutCss = css`
-  padding: 120px 0;
   display: flex;
   flex-direction: column;
   align-items: center;
-  gap: 36px;
-  background-color: white;
-
-  ${mediaQuery('mobile')} {
-    padding: 0;
-    background-color: ${theme.colors.lightGray};
-  }
+  width: 100%;
+  background-color: ${theme.colors.black};
 `;
 
+const introCss = {
+  headline: css`
+    ${theme.typosV2.pretendard.bold32}
+    line-height: 150%;
+    color: ${theme.colors.white};
+
+    ${mediaQuery('mobile')} {
+      ${theme.typosV2.pretendard.bold28}
+      line-height: 150%;
+    }
+  `,
+  description: css`
+    ${theme.typosV2.pretendard.semibold20}
+    line-height: 150%;
+    margin-top: 42px;
+    text-align: center;
+    color: ${theme.colors.white};
+
+    ${mediaQuery('mobile')} {
+      ${theme.typosV2.pretendard.semibold18}
+    }
+  `,
+};
+
 const supportContainerCss = css`
+  margin-top: 88px;
   display: grid;
   width: 100%;
   max-width: 960px;
@@ -52,32 +73,11 @@ const supportContainerCss = css`
   column-gap: 12px;
   justify-items: center;
   justify-content: center;
+  padding: 0 20px;
 
-  ${mediaQuery('tablet')} {
+  ${mediaQuery('mobile')} {
     grid-template-columns: repeat(2, 1fr);
-    padding: 0 32px;
   }
-
-  @media (max-width: 600px) {
-    grid-template-columns: repeat(1, 1fr);
-    padding: 0 16px;
-  }
-`;
-
-const titleContainerCss = css`
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-`;
-
-const titleCss = (theme: Theme) => css`
-  ${theme.typosV2.pretendard.semibold20}
-`;
-
-const subTitleCss = (theme: Theme) => css`
-  text-align: center;
-  ${theme.typosV2.pretendard.semibold16}
-  color: #555;
 `;
 
 const imageContainerCss = css`
