+++ src/features/About/sections/AboutSupportsSection.tsx
@@ -1,32 +1,14 @@
-import Image from 'next/image';
 import { css } from '@emotion/react';
 
-import { SupportThumbnail } from '~/components/Supports/SupportThumbnail';
-import { SUPPORTS } from '~/constant/supports';
-import { useCheckWindowSize } from '~/hooks/useCheckWindowSize';
+import { Supports } from '~/components/Supports';
 import { colors } from '~/styles/colors';
 import { mediaQuery } from '~/styles/media';
 import { theme } from '~/styles/theme';
 
 export const AboutSupportsSection = () => {
-  const { isTargetSize: isMobileSize } = useCheckWindowSize('mobile');
-
   return (
     <div css={[layoutCss]}>
-      <h1 css={introCss.headline}>후원사</h1>
-      <p css={introCss.description}>
-        디프만은 IT비영리단체로 후원을 통해
-        {isMobileSize && <br />}더 많은 교육 기회에 도움을 받고 있습니다.
-      </p>
-
-      <ul css={supportContainerCss}>
-        {SUPPORTS.map(support => (
-          <SupportThumbnail key={support.title} {...support} />
-        ))}
-        <div css={imageContainerCss}>
-          <Image src="/images/support/next.png" alt="what's next?" fill quality={100} />
-        </div>
-      </ul>
+      <Supports />
 
       <button css={supportButtonCss}>후원 문의하기</button>
     </div>
@@ -38,60 +20,8 @@ const layoutCss = css`
   display: flex;
   flex-direction: column;
   align-items: center;
+  width: 100%;
   background-color: ${theme.colors.black};
-`;
-
-const introCss = {
-  headline: css`
-    ${theme.typosV2.pretendard.bold32}
-    line-height: 150%;
-    color: ${theme.colors.white};
-
-    ${mediaQuery('mobile')} {
-      ${theme.typosV2.pretendard.bold28}
-      line-height: 150%;
-    }
-  `,
-  description: css`
-    ${theme.typosV2.pretendard.semibold20}
-    line-height: 150%;
-    margin-top: 42px;
-    text-align: center;
-    color: ${theme.colors.white};
-
-    ${mediaQuery('mobile')} {
-      ${theme.typosV2.pretendard.semibold18}
-    }
-  `,
-};
-
-const supportContainerCss = css`
-  margin-top: 88px;
-  display: grid;
-  width: 100%;
-  max-width: 960px;
-  grid-template-columns: repeat(3, 1fr);
-  row-gap: 20px;
-  column-gap: 12px;
-  justify-items: center;
-  justify-content: center;
-  padding: 0 20px;
-
-  ${mediaQuery('mobile')} {
-    grid-template-columns: repeat(2, 1fr);
-  }
-`;
-
-const imageContainerCss = css`
-  position: relative;
-  width: 100%;
-  height: 208px;
-
-  ${mediaQuery('mobile')} {
-    max-width: 450px;
-    padding: 18px;
-    height: 180px;
-  }
 `;
 
 const supportButtonCss = css`
